requirements.txt
Flask>=2.0
pillow
pytesseract
opencv-python
sqlite3-binary
pandas